Move ComputedUrl into their impl mods.
